+++ cli.ts
@@ -44,7 +44,9 @@ import { health } from './lib/commands/services/health.js';
 import { locale } from './lib/commands/services/locale.js';
 import { messaging } from './lib/commands/services/messaging.js';
 import { migrations } from './lib/commands/services/migrations.js';
+import { organization } from './lib/commands/services/organization.js';
 import { organizations } from './lib/commands/services/organizations.js';
+import { presences } from './lib/commands/services/presences.js';
 import { project } from './lib/commands/services/project.js';
 import { projects } from './lib/commands/services/projects.js';
 import { proxy } from './lib/commands/services/proxy.js';
@@ -214,7 +216,9 @@ if (process.argv.includes('-v') || process.argv.includes('--version')) {
             .addCommand(locale)
             .addCommand(messaging)
             .addCommand(migrations)
+            .addCommand(organization)
             .addCommand(organizations)
+            .addCommand(presences)
             .addCommand(project)
             .addCommand(projects)
             .addCommand(proxy)